Refactor StoreInfoMaker package
