+++ src/main/java/com/neconico/neconico/controller/users/UserController.java
@@ -3,7 +3,7 @@ package com.neconico.neconico.controller.users;
 import com.neconico.neconico.config.web.LoginUser;
 import com.neconico.neconico.dto.users.*;
 import com.neconico.neconico.service.store.StoreInfoService;
-import com.neconico.neconico.service.store.provider.StoreInfoMaker;
+import com.neconico.neconico.Maker.StoreInfoMaker;
 import com.neconico.neconico.service.users.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
